First version of aionfix.
This fixes extremely low FPS in WINE when tap-tabbing targets in some legacy
Aion game clients (used by some private servers). The issue seems to be that
the client during tab-tabbing issues an excessive number of calls to
NtReadVirtualMemory() which in WINE is implemented in WINE server using
ptrace attack/read/detach to read the client memory which causes massive
stalls.

The workaround presented here basically ptraces all its children processes,
which effectively causes the WINE server's ptrace() call to fail (there can be
only one ptrace master process). This consequentially causes the
NtReadVirtualMemory() call from the client to fail, but it doesn't seem to
cause too much harm and boosts the FPS to _normal_ levels.

Presumably the reason why the client is issuing a lot of
NtReadVirtualMemory() calls is because of some anti-cheat code (it doesn't
make sense to use that function on itself otherwise).

strace could be used instead, but since it traces all system calls it also
induces noticable jitter.